mmor: Cleanup Windows module enumeration code
- Added more details to module_to_object error message to capture
  which module lookup failed
- Try to use the correct module enumeration flag value depending
  on architecture (see below)

In a project that uses mmor in a Windows 32-bit context, one of
the GetModuleInformation calls fails with an "invalid handle" error.
I think when we ask for a list of Windows modules, it is giving us
*all* of them including handles to 64-bit DLLs (I do not understand
how 64-bit DLLs are loaded in a 32-bit program... but I see them in
the process's vmmap, so idk).

The handle it is consistently failing on is a 64-bit DLL (shell32.dll).
The Windows EnumProcessModulesEx documentation claims you can filter
modules, so I tried only asking for 32-bit modules. After doing that,
I get the same error on the same module. The function's documentation
has this fun statement:

> This function is intended primarily for 64-bit applications. If the
> function is called by a 32-bit application running under WOW64, the
> dwFilterFlag option is ignored and the function provides the same
> results as the EnumProcessModules function.

And the documentation for the EnumProcessModules function says:

> If this function is called from a 32-bit application running on
> WOW64, it can only enumerate the modules of a 32-bit process.

So I guess the documentation is wrong?